*Implemented BossList Saved variable:
This functions by adding an NPC's name to the list, followed by the full path to the song you want to play. This is intened for tech-savvy persons, and will not be supported.

To add your target:
/run CombatMusic_BossList[UnitName('target')] = "SomeSong.mp3"

If you want to add for anything else, just replace "UnitName('target')" with the NPC's exact, full name, in quotation marks (I.E. "Heroic Training Dummy")

*Reset Dialog now warns players that thier boss list will be reset on pressing Okay.

*Startup settings checks should work properly now. (Attempted fix for the no-settings bug.)

*Fixed Addon Startup messages, They'll show up in the chat frame again.