Tagging as release_v3.12
Major changes inclue:
Adding a fadeout function for currently playing music. Set /cm fade to the number of seconds you want the music to fade out across to enable this feature.

Added a SavedVariables checking. It now checks to see if your SavedVars are up to date, and resets them if they are not.