Added a new feature where you can now add bosslist entries through the slash command interface with /cm bosslist add
This will pop up a series of dialogs that will ask you for the NPC's name and the path to the music to play. make sure to escape the backslashes in the path with double ones! I won't help you fix errors caused by BossList adding!

Reset your CombatMusic_BossList if this happens!