Bump SavedVars to Version 2. I am too lazy to add code to convert, so your settings are going to be reset, sorry.
Added a few new variables to the format, and it's a lot neater now. One of the new features is the ability to have another file for your level up sound, DING.mp3 is what the program will be looking for, and you will have to tell it that you want to use it instead of the default file.