New target checking works New method of checking targets.
Fixed a bug where if you were to engage a 'boss' that was not in combat or on the bosslist during a fadeout, it would not play boss music.

Added 'DEV' revision if project_revision can't be determined.

Debug messages will now prefer ChatFrame4 if it's visible, otherwise they will default to ChatFrame1